fixed Bug 43404 - JSS - Composite Elements - properties are not saved after changing fixed Bug 43403 - JSS - Composite Elements - incorrect section by default
git-svn-id: http://code.jaspersoft.com/svn/repos/jaspersoftstudio/trunk@5985 6833c83e-4fd0-4c68-a3df-283c6148db30

+++ com.jaspersoft.studio/src/com/jaspersoft/studio/editor/tools/CompositeElementManager.java
@@ -480,11 +480,18 @@ public class CompositeElementManager {
 		File storage = ConfigurationManager.getStorage(ELEMENTS_STORAGE_KEY);	
 		try{
 			String oldName = oldElement.getName();
-			File oldIconSmall = new File(oldElement.getIconPathSmall()); //$NON-NLS-1$
-			oldIconSmall.delete();
-			File oldIconBig = new File(oldElement.getIconPathBig()); //$NON-NLS-1$
-			oldIconBig.delete();
 			
+			//delete the old images if present
+			if (oldElement.getIconPathSmall() != null){
+				File oldIconSmall = new File(oldElement.getIconPathSmall()); //$NON-NLS-1$
+				oldIconSmall.delete();
+			}
+			
+			if (oldElement.getIconPathBig() != null){
+				File oldIconBig = new File(oldElement.getIconPathBig()); //$NON-NLS-1$
+				oldIconBig.delete();
+			}
+				
 			//Write the report on disk
 			String oldReportName = oldName + COMPOSITE_ELEMENT_EXTENSION;
 			String newReportName = newName + COMPOSITE_ELEMENT_EXTENSION;
